convert regular polygon tests to vitest
